move resolvers to seperate files and merge them in root resolver
